+++ responder.go
@@ -10,23 +10,23 @@ type Responder interface {
 	Preload()
 	Setup()
 	Close()
-	Update(dt float32)
-	Mouse(x, y float32, action Action)
-	Scroll(amount float32)
+	Update(dt float64)
+	Mouse(x, y int, action Action)
+	Scroll(amount float64)
 	Key(key Key, modifier Modifier, action Action)
 	Type(char rune)
 }
 
 type Game struct{}
 
-func (g *Game) Preload()                          {}
-func (g *Game) Setup()                            {}
-func (g *Game) Close()                            {}
-func (g *Game) Update(dt float32)                 {}
-func (g *Game) Render()                           {}
-func (g *Game) Resize(w, h int)                   {}
-func (g *Game) Mouse(x, y float32, action Action) {}
-func (g *Game) Scroll(amount float32)             {}
+func (g *Game) Preload()                      {}
+func (g *Game) Setup()                        {}
+func (g *Game) Close()                        {}
+func (g *Game) Update(dt float64)             {}
+func (g *Game) Render()                       {}
+func (g *Game) Resize(w, h int)               {}
+func (g *Game) Mouse(x, y int, action Action) {}
+func (g *Game) Scroll(amount float64)         {}
 func (g *Game) Key(key Key, modifier Modifier, action Action) {
 	if key == Escape {
 		Exit()
